+++ src/model/mod.rs
@@ -2,7 +2,4 @@ pub mod cst_doc;
 pub mod cst_edit;
 pub mod cst_project;
 pub mod document;
-pub mod fragment;
 pub mod node;
-pub mod project;
-pub mod toml_doc;